+++ src/js/components/RepositoriesTabUI.js
@@ -1,0 +1,104 @@
+import { Link } from "react-router";
+import React from "react";
+
+import AddRepositoryFormModal from "#SRC/js/components/modals/AddRepositoryFormModal";
+import Breadcrumb from "#SRC/js/components/Breadcrumb";
+import BreadcrumbTextContent from "#SRC/js/components/BreadcrumbTextContent";
+import FilterBar from "#SRC/js/components/FilterBar";
+import FilterInputText from "#SRC/js/components/FilterInputText";
+import Loader from "#SRC/js/components/Loader";
+import Page from "#SRC/js/components/Page";
+import RepositoriesTable from "#SRC/js/components/RepositoriesTable";
+import RequestErrorMsg from "#SRC/js/components/RequestErrorMsg";
+
+const RepositoriesBreadcrumbs = addButton => {
+  const crumbs = [
+    <Breadcrumb key={-1} title="Repositories">
+      <BreadcrumbTextContent>
+        <Link to="/settings/repositories">Package Repositories</Link>
+      </BreadcrumbTextContent>
+    </Breadcrumb>
+  ];
+
+  return (
+    <Page.Header.Breadcrumbs
+      iconID="settings"
+      breadcrumbs={crumbs}
+      addButton={addButton}
+    />
+  );
+};
+
+const METHODS_TO_BIND = ["handleCloseAddRepository", "handleOpenAddRepository"];
+
+export default class RepositoriesTabUI extends React.Component {
+  constructor() {
+    super();
+
+    this.state = {
+      addRepositoryModalOpen: false
+    };
+
+    METHODS_TO_BIND.forEach(method => {
+      this[method] = this[method].bind(this);
+    });
+  }
+
+  handleCloseAddRepository() {
+    this.setState({ addRepositoryModalOpen: false });
+  }
+
+  handleOpenAddRepository() {
+    this.setState({ addRepositoryModalOpen: true });
+  }
+
+  getContent() {
+    const { addRepositoryModalOpen } = this.state;
+
+    const { repositories, isLoading, hasError, searchString } = this.props;
+
+    // console.log("render", this.props, this.state);
+
+    if (hasError) {
+      return <RequestErrorMsg />;
+    }
+
+    if (isLoading || typeof repositories !== "object") {
+      return <Loader />;
+    }
+
+    return (
+      <div>
+        <FilterBar>
+          <FilterInputText
+            className="flush-bottom"
+            placeholder="Search"
+            searchString={searchString}
+            handleFilterChange={this.props.onSearch}
+          />
+        </FilterBar>
+        <RepositoriesTable repositories={repositories} filter={searchString} />
+        <AddRepositoryFormModal
+          numberOfRepositories={repositories.getItems().length}
+          open={addRepositoryModalOpen}
+          onClose={this.handleCloseAddRepository}
+        />
+      </div>
+    );
+  }
+
+  render() {
+    return (
+      <Page>
+        <Page.Header
+          addButton={{
+            onItemSelect: this.handleOpenAddRepository,
+            label: "Add Repository"
+          }}
+          breadcrumbs={<RepositoriesBreadcrumbs />}
+        />
+        {this.getContent()}
+      </Page>
+    );
+  }
+}
